+++ CVR_Code/src/main/java/frc/robot/subsystems/DriveSubsystem.java
@@ -187,29 +187,4 @@ public class DriveSubsystem extends SubsystemBase {
 	public void calibrate() {
 		ahrs.calibrate();
 	}
-
-	public Command getBalanceCommand(double minVoltage, double maxVoltage, double delayThresholdDegPerS) {
-		final double maxPitch = 20.0;
-		MathUtil.SpeedGetter speedGetter = new MathUtil.SpeedGetter(() -> {return getPitch();});
-
-		Supplier<Double> fwd = () -> {
-			return speedGetter.get() > delayThresholdDegPerS ? 0 :
-				MathUtil.scaleMagnitude(getPitch(), 0, maxPitch, minVoltage, maxVoltage, 1.5);
-		};
-
-		
-		return run(() -> arcadeDrive(fwd.get(), 0));
-	}
-
-    public Command getTimedDrive(long timeMS, double power) {
-        final class Timer {
-            long setPoint;
-            public Timer(long timeMS) { this.setPoint = System.currentTimeMillis() + timeMS; }
-            public boolean isFinished() { return System.currentTimeMillis() >= setPoint; }
-        };
-        
-        Timer timer = new Timer(timeMS);
-        return run(() -> arcadeDrive(-power, 0))
-            .until(timer::isFinished).andThen(runOnce(()-> arcadeDrive(0, 0)));
-    }
 }
